Fixed merge problems and updated tests
