+++ lib/aurora-sdk/interpret.rb
@@ -1,8 +1,6 @@
-require 'json'
-
 module Aurora
     class Interpret
-        attr_reader :text, :intent, :entities
+        attr_accessor :text, :intent, :entities
 
         def initialize(text, intent, entities)
             @text = text
